feature/mapstruct: add dependency mapstruct, create class BrandMapper and deleted class old mapper

+++ src/main/java/com/phoneshop/org/model/entity/Model.java
@@ -23,7 +23,7 @@ public class Model {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "model_id")
-    private Long modelId;
+    private Integer modelId;
     
     @Column(name = "model_name")
     private String modelName;
